feat: allow PSR-7 v2
Bumps the psr/http-message dependency to allow v2.

Test changes are due to the fact that a number of mocks were not returning the correct value, and with RTH defined in v2, these were now raising errors.
With the changes, they are now more correct, and test assertions continued to pass without issue.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>